requirements.txt
# RAG pipeline — pinned for reproducibility
chromadb==0.5.23
sentence-transformers==3.3.1
torch>=2.2.0
faster-whisper==1.1.0
python-pptx==1.0.2
httpx==0.28.1
python-dotenv==1.0.1
numpy>=1.26.0
